+++ support/cas-server-support-ws-idp/src/main/java/org/apereo/cas/config/CoreWsSecurityIdentityProviderConfiguration.java
@@ -125,7 +125,7 @@ class CoreWsSecurityIdentityProviderConfiguration {
             return plan -> {
                 LOGGER.debug("Initializing WS Federation callback service [{}]", wsFederationCallbackService);
                 val service = new CasRegisteredService();
-                service.setId(RandomUtils.nextLong());
+                service.setId(RandomUtils.nextInt());
                 service.setEvaluationOrder(Ordered.HIGHEST_PRECEDENCE);
                 service.setName(service.getClass().getSimpleName());
                 service.setDescription("WS-Federation Authentication Request");
